Se agregaron rutas y sus consultas

+++ Eleccion/modelo.py
@@ -1,6 +1,6 @@
 from sqlalchemy.sql.schema import ForeignKey
 import db
-from sqlalchemy import Column, Integer, Boolean, String
+from sqlalchemy import Column, Integer, Boolean, String, BigInteger
 from pydantic import BaseModel
 
 ## MODELO RELACIONADO CON LA BD
@@ -8,8 +8,8 @@ class Eleccion(db.Base):
     """Este modelo define los atributos de la tabla eleccion y sus tipos de dato"""
     __tablename__ = "eleccion"
     codigo = Column("codigo", String(255), primary_key=True, unique=True, index=True)
-    fecha_inicio = Column("fecha_inicio", Integer, unique=True)
-    fecha_fin = Column("fecha_fin", Integer, unique=True)
+    fecha_inicio = Column("fecha_inicio", BigInteger, unique=True)
+    fecha_fin = Column("fecha_fin", BigInteger, unique=True)
     nombre = Column("nombre", String(255))
     descripcion = Column("descripcion", String(255))
     cedula_candidato = Column(String(255), ForeignKey("candidato.cedula"))
